Adding the ability to load Extension functionality based on a Signal.
While Extensions are still instantiated when they are first attached,
there internal setup code (init()) is now not called until an event is
triggered.

This gives us the ability to only init() an Extension when it is really
required.

This also opens us up to the idea of having a "Module Loader" of sorts
that will only alter the Routers routes when required. This would make
an excellent plugin.

Closes #70